+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add extension fields support for PHP
 - Add extension fields support for Ruby
 - Add extension fields support for Bash
+- Add extension fields support for Scala
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers